backend/requirements.txt
Flask==2.3.2
Authlib==1.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
Flask-Cors==4.0.1
requests==2.31.0
